+++ db/migrate/20180730191017_add_locked_to_topics.rb
@@ -1,0 +1,6 @@
+class AddLockedToTopics < ActiveRecord::Migration[5.2]
+  def change
+    add_column :topics, :locked, :boolean, null: false, default: false
+    add_index :topics, :locked
+  end
+end