+++ CHANGELOG.md
@@ -27,6 +27,7 @@
   - Simplified user export
 - **Forum Changes**
   - Added a prominent forum search to the forum index
+  - Topics now auto-lock after two months of inactivity
 - **Search Changes**
   - Member profile previews are displayed when searching usernames
   - Added customized search results